Convert ServerNotReadyQuery to use mustache and add memory details to IndexQuery
