+++ src/com/sun/tools/hat/internal/server/ServerNotReadyQuery.java
@@ -34,48 +34,19 @@ package com.sun.tools.hat.internal.server;
 
 import com.sun.tools.hat.internal.parser.LoadProgress;
 
-import java.lang.management.ManagementFactory;
-import java.lang.management.MemoryUsage;
-
-class ServerNotReadyQuery extends QueryHandler {
-    private static final String MB = "MB";
-    private static final String GB = "GB";
+class ServerNotReadyQuery extends MustacheQueryHandler {
     private final LoadProgress loadProgress;
+    private MemoryUsageDetails memoryUsage = new MemoryUsageDetails();
 
     public ServerNotReadyQuery(LoadProgress loadProgress) {
         this.loadProgress = loadProgress;
     }
 
-    @Override
-    public void run() {
-        startHtml("Server Not Ready");
-        printMemoryUsage();
-        loadProgress.each(this::printProgress);
-        out.println("<meta http-equiv=\"refresh\" content=\"1\" />");
-        endHtml();
+    public LoadProgress getLoadProgress() {
+        return loadProgress;
     }
 
-    private void printProgress(LoadProgress.ProgressElement progress) {
-        out.println("<p>");
-        println(progress.getLoadString());
-        out.println("</p>");
-    }
-
-    private void printMemoryUsage() {
-        MemoryUsage memory = ManagementFactory.getMemoryMXBean().getHeapMemoryUsage();
-        String size = MB;
-        double used = (double) memory.getUsed() / 1024.0 / 1024.0;
-        double total = (double) memory.getMax() / 1024.0 / 1024.0;
-        double percentUsed = 100.0 * used / total;
-
-        if (total >= 1024.0) {
-            size = GB;
-            used /= 1024.0;
-            total /= 1024.0;
-        }
-
-        out.println("<p>");
-        out.println(String.format("<b>Heap Utilization:</b> %1.2f%s / %1.2f%s (%1.1f%%)", used, size, total, size, percentUsed));
-        out.println("</p>");
+    public MemoryUsageDetails getMemoryUsage() {
+        return memoryUsage;
     }
 }
